Added shuffle on copycat data
When the NPD images are labeled, there is a very little chance to be ordered by classes (i.e., images ordered by their classes inside text file). But I am adding shuffle on dataloader to avoid problems if the user sort the images per class or do something similar to it.